dev-556: refactor index test utils
