+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2021-03-09T16:23:06Z",
+  "generated_at": "2021-08-02T18:21:49Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -92,14 +92,14 @@
         "hashed_secret": "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 31,
+        "line_number": 23,
         "type": "Basic Auth Credentials"
       },
       {
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 85,
+        "line_number": 222,
         "type": "Secret Keyword"
       }
     ],
